+++ src/main/java/com/smart/event/event_management_system/service/BookingService.java
@@ -7,6 +7,6 @@ import com.smart.event.event_management_system.entity.Booking;
 public interface BookingService {
     Booking createBooking(Long userId, Long eventId, int tickets);
     List<Booking> getBookingsByUser(Long userId);
-    List<Booking> getBookingsByEvent(Integer eventId);
-    void cancelBooking(Integer bookingId);
+    List<Booking> getBookingsByEvent(Long eventId);
+    void cancelBooking(Long bookingId);
 }
